+++ src/commands/publish.ts
@@ -5,7 +5,10 @@ import SiteStore from "../utils/site-store";
 export default async () => {
   const activeEditor = vscode.window.activeTextEditor;
   if (activeEditor) {
-    await vscode.commands.executeCommand("vscode-extension-halo.upload-images");
+    const siteStore = new SiteStore();
+    const service = new HaloService(siteStore.getDefaultSite());
+
+    await service.uploadImages();
 
     vscode.window.withProgress(
       {
@@ -14,8 +17,6 @@ export default async () => {
         cancellable: false,
       },
       async () => {
-        const siteStore = new SiteStore();
-        const service = new HaloService(siteStore.getDefaultSite());
         service.publishPost();
       }
     );