+++ src/commands/upload-images.ts
@@ -3,16 +3,7 @@ import HaloService from "../service";
 import SiteStore from "../utils/site-store";
 
 export default async () => {
-  await vscode.window.withProgress(
-    {
-      location: vscode.ProgressLocation.Notification,
-      title: vscode.l10n.t("Uploading images..."),
-      cancellable: false,
-    },
-    async () => {
-      const siteStore = new SiteStore();
-      const service = new HaloService(siteStore.getDefaultSite());
-      service.uploadImages();
-    }
-  );
+  const siteStore = new SiteStore();
+  const service = new HaloService(siteStore.getDefaultSite());
+  service.uploadImages();
 };