+++ src/commands/update.ts
@@ -1,8 +1,10 @@
 import HaloService from "../service";
 import SiteStore from "../utils/site-store";
+import * as vscode from "vscode";
 
 export default async () => {
   const siteStore = new SiteStore();
   const service = new HaloService(siteStore.getDefaultSite());
-  service.uploadImages();
+  service.updatePost();
+  vscode.window.showInformationMessage(vscode.l10n.t("Post updated"));
 };