+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 [unreleased]
 
 - Exclude deleted posts when pulling posts
+- Exclude hidden attachment group when setup
 
 ## 1.0.0