+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Exclude deleted posts when pulling posts
 - Exclude hidden attachment group when setup
+- Support batch pulling posts
 
 ## 1.0.0