+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add publish parameters to Front-matter
 - Optimize the logic of picture upload
 - Support setting categories, tags and title
+- Support updating post from halo remote
 
 ## 1.0.0